+++ seeders/RolesSeeder.js
@@ -2,22 +2,27 @@ module.exports = {
 	rolesSeeder: runRolesSeeder()
 }
 
-function runRolesSeeder() {
+async function runRolesSeeder() {
 	const Role = require('../models/Role');
 	const mongoose = require('mongoose');
 	
 	const ATLAS_URI = require('../config/connection').mongoURI;
 	const uri = ATLAS_URI || 'mongodb://localhost:27017/capstone3_db';
-	// const uri = 'mongodb://localhost:27017/capstone3_db';
-	mongoose.connect(uri, {
+	await mongoose.connect(uri, {
 		useNewUrlParser: true,
 		useUnifiedTopology: true,
 		useCreateIndex: true,
 		useFindAndModify: false
 	})
-	.then(() => {
+	.then(async () => {
 		console.log('RolesSeeder connected')
-		mongoose.connection.db.dropCollection('roles');
+		const docCount = await Role.estimatedDocumentCount();
+		if(docCount > 0) {
+			mongoose.connection.db.dropCollection('roles');
+			console.log("'roles' collection successfully dropped")
+		} else {
+			console.log("'roles' collection is empty")
+		}
 	})
 	.catch((err) => console.log(err));
